work in progress Power dissipation

+++ PD.cpp
@@ -8,54 +8,54 @@ T::PD::PD(Values values)
 void T::PD::AddValues(Values values) { _values = values; }
 T::PD::Values T::PD::ShowValues() { return _values; }
 
-int T::PD::CalcTcon()
+double T::PD::CalcTcon()
 {
 	if (_values.tcon == 0) {
 		_values.tcon = _values.tri + _values.tfv;
 	}
 	return _values.tcon;
 }
-int T::PD::CalcTcoff()
+double T::PD::CalcTcoff()
 {
 	if (_values.tcoff == 0) {
 		_values.tcoff = _values.trv + _values.tfi;
 	}
 	return _values.tcoff;
 }
-int T::PD::CalcWcon()
+double T::PD::CalcWcon()
 {
 	if (_values.Wcon == 0) {
 		_values.Wcon = 0.5 * _values.VDC * _values.Io * CalcTcon();
 	}
 	return _values.Wcon;
 }
-int T::PD::CalcWcoff()
+double T::PD::CalcWcoff()
 {
 	if (_values.Wcoff == 0) {
 		_values.Wcoff = 0.5 * _values.VDC * _values.Io * CalcTcoff();
 	}
 	return _values.Wcoff;
 }
-int T::PD::CalcPs()
+double T::PD::CalcPs()
 {
 	if (_values.Ps == 0) {
 		_values.Ps = (CalcWcon() + CalcWcoff()) * _values.fs;
 	}
 	return _values.Ps;
 }
-int T::PD::CalcWon()
+double T::PD::CalcWon()
 {
 	if (_values.Won == 0) { _values.Won = _values.Von * _values.Io * CalcTcon(); }
 	return _values.Won;
 }
-int T::PD::CalcPon()
+double T::PD::CalcPon()
 {
 	if (_values.Pon == 0) {
 		_values.Pon = CalcWon() * _values.fs;
 	}
 	return _values.Pon;
 }
-int T::PD::CalcPdis()
+double T::PD::CalcPdis()
 {
 	_values.Pdis = CalcPs() + CalcPon();
 	return _values.Pdis;
